중복채점 방지 및 소수점 출력 제거

+++ src/main/java/swm_nm/morandi/member/dto/GraphDto.java
@@ -11,5 +11,5 @@ import java.util.Map;
 @Builder
 public class GraphDto {
     private String algorithmName;
-    private Double solvedRate;
+    private Long solvedRate;
 }
